Fully abstract IPC file. Fix some bad param passing.
Still pretty rapid internal development, sprry for bad commit message.

+++ src/orderBuilds.sh
@@ -46,12 +46,9 @@ LUNCH_COMMAND="lunch ${1}_${TARGET}-userdebug"
 # Needed to allow for AOKP, and anyone else who gets cute.
 DEFAULT_LUNCH_COMMAND="lunch aosp_$TARGET-userdebug"
 
-
 # If you want the whole rom, comment this line and uncomment the next.
-[[ "$TARGET_VERSION" == "" ]] && BUILD_COMMAND=$PATCH_SCRIPTS_LOC/src/makeOPDFiles.sh
+BUILD_COMMAND=$PATCH_SCRIPTS_LOC/src/makeOPDFiles.sh
 #BUILD_COMMAND="make otapackage"
-
-JAR_OUT="$1"-"$TARGET_VERSION"
 
 print_error() {
      echo ""
@@ -242,10 +239,15 @@ cd $ANDROID_HOME
 # if OpD patches are applied, don't repo sync.
 if [[ ! -f $LOCK ]]; then
      # get of legitimate as well as hacky manifests from old builds.
-     rm -rf .repo/manifests manifests.xml
-     rm -rf .repo/local_manifests local_manifests.xml
-     $REPO_SYNC_COMMAND
-     repo sync -j${JOBS} -f
+     current=$(cat $IPC)
+     current_rom=${current%-*}
+     # remove old manifests and repo init if switching rom types
+     if ( "$1" != current ); then
+          rm -rf .repo/manifests manifests.xml
+          rm -rf .repo/local_manifests local_manifests.xml
+          $REPO_SYNC_COMMAND
+          repo sync -j${JOBS} -f
+     fi
      . build/envsetup.sh
      $LUNCH_COMMAND
      # for roomservice and local_manifest pickups
